Select multiple bag files from open dialog.
This allows a user to select multiple files to load at once from the
read bags dialog.  Performance isn't so hot right now because it
spawns a new thread for each file and they all just hammer the disk.
May change to a single bag reader with a queue later.