Bug fix & improvements to collect_rosout_bags.
I ran this utility on a set of 3736 bag files and it crashed on bag
3686 bags after running for 20 hours.  There was an uncaught exception
related to log messages with a timestamp of 0, which can happen when a
node is publishing while simtime is enabled, but before a clock is
being published.  This commit fixes that case.

This commit also adds a log message at the start of each bag file that
we can use to identify breaks in the bag file.  It also prints out the
total message count as we're going along.