Restoring font selection features.
This was changed to make font selection independent between
windows. They were originally made to be shared because we didn't have
the nifty save/restore settings functionality, so it was important to
be able to change them globally.  Now that we have that, it's easy
enough to save it once and then restart the console.  Removing the
behavior simplifies the master/window interface and makes the font
setting consistent with how all other settings are saved now.